wip uv

+++ bin/install.ps1
@@ -1,7 +1,6 @@
 $OLD_CWD = (Get-Location).Path
 
 $SRC_PATH = "$env:LOCALAPPDATA\Programs\socketsupply\src"
-$LIB_PATH = "$env:LOCALAPPDATA\Programs\socketsupply\lib"
 $WORKING_PATH = $OLD_CWD
 
 #
@@ -13,19 +12,21 @@ Function Build {
   $BUILD_TIME = [int] (New-TimeSpan -Start (Get-Date "01/01/1970") -End (Get-Date)).TotalSeconds
   $BUILD_PATH = "$WORKING_PATH\build"
 
-  (git clone -q --depth=1 git@github.com:libuv/libuv.git $BUILD_PATH) > $null
-  (New-Item -ItemType Directory -Force -Path "$BUILD_PATH\lib") > $null
-  Write-Output "ok - cloned libuv $BUILD_PATH"
+  if (-not (Test-Path -Path $BUILD_PATH -PathType Container)) {
+    (git clone -q --depth=1 git@github.com:libuv/libuv.git $BUILD_PATH) > $null
+    (New-Item -ItemType Directory -Force -Path "$BUILD_PATH\lib") > $null
+    Write-Output "ok - cloned libuv $BUILD_PATH"
 
-  cd "$BUILD_PATH\lib"
-  Write-Output "# bulding libuv..."
-  (cmake ..) > $null
-  cd "$WORKING_PATH"
-  (cmake --build "$BUILD_PATH\lib" --config Release) > $null
-  Write-Output "ok - built libuv"
+    cd "$BUILD_PATH\lib"
+    Write-Output "# bulding libuv..."
+    (cmake ..) > $null
+    cd "$WORKING_PATH"
+    (cmake --build "$BUILD_PATH\lib" --config Release) > $null
+    Write-Output "ok - built libuv"
 
-  Copy-Item -Path "$BUILD_PATH\include\*" -Destination $SRC_PATH -Recurse -Container
-  Copy-Item $BUILD_PATH\lib\Release\uv_a.lib -Destination $SRC_PATH\uv
+    Copy-Item -Path "$BUILD_PATH\include\*" -Destination $SRC_PATH -Recurse -Container
+    Copy-Item $BUILD_PATH\lib\Release\uv_a.lib -Destination $SRC_PATH\uv
+  }
 
   Write-Output "# compiling the build tool..."
   clang++ src\cli.cc -o $WORKING_PATH\bin\ssc.exe -std=c++2a -DBUILD_TIME="$($BUILD_TIME)" -DVERSION_HASH="$($VERSION_HASH)" -DVERSION="$($VERSION)"
@@ -73,8 +74,8 @@ Function Build {
 #
 Function Install-Files {
   Copy-Item $WORKING_PATH\bin\ssc.exe -Destination $SRC_PATH
-  Copy-Item -Path "$WORKING_PATH\src\*" -Destination $SRC_PATH -Recurse
-  Copy-Item -Path "$WORKING_PATH\src\win64\*" -Destination $SRC_PATH -Recurse
+  Copy-Item -Path "$WORKING_PATH\src\*" -Destination $SRC_PATH -Recurse -Force
+  Copy-Item -Path "$WORKING_PATH\src\win64\*" -Destination $SRC_PATH -Recurse -Force
   Write-Output "ok - installed files to '$SRC_PATH'."
 }
 
@@ -109,7 +110,7 @@ if ($? -ne 1) {
   iex ((New-Object System.Net.WebClient).DownloadString('https://community.chocolatey.org/install.ps1'))
 }
 
-(Get-Command clang++.exe) > $null
+(Get-Command "clang++.exe") > $null
 
 if ($? -ne 1) {
   choco install llvm
@@ -120,7 +121,7 @@ if ($? -ne 1) {
   }
 }
 
-(Get-Command cmake.exe) > $null
+(Get-Command "cmake.exe") > $null
 
 if ($? -ne 1) {
   choco install cmake --installargs 'ADD_CMAKE_TO_PATH=System'
@@ -131,7 +132,7 @@ if ($? -ne 1) {
   }
 }
 
-(Get-Command git) > $null
+(Get-Command "git.exe") > $null
 
 if ($? -ne 1) {
   choco install git
@@ -147,21 +148,10 @@ if ($args[0] -eq "update") {
   Exit 0
 }
 
-if (Test-Path -Path $SRC_PATH) {
-  Remove-Item -Recurse -Force $SRC_PATH
-  Write-Output "ok - cleaned $SRC_PATH"
+if (-not (Test-Path -Path $SRC_PATH)) {
+  (New-Item -ItemType Directory -Path $SRC_PATH) > $null
+  Write-Output "ok - created $SRC_PATH"
 }
-
-if (Test-Path -Path $LIB_PATH) {
-  Remove-Item -Recurse -Force $LIB_PATH
-  Write-Output "ok - cleaned $LIB_PATH"
-}
-
-(New-Item -ItemType Directory -Force -Path $SRC_PATH) > $null
-Write-Output "ok - created $SRC_PATH"
-
-(New-Item -ItemType Directory -Force -Path $LIB_PATH) > $null
-Write-Output "ok - created $LIB_PATH"
 
 Write-Output "# working path set to $WORKING_PATH"
 cd $WORKING_PATH
